+++ engine/include/reader/fmt.h
@@ -1,0 +1,12 @@
+/* File Format Enumeration Header */
+
+#ifndef VMCU_FMT_H
+#define VMCU_FMT_H
+
+typedef enum {
+
+    VMCU_FMT_IHEX
+
+} VMCU_FMT;
+
+#endif